+++ test/lib/folio/devise/crossdomain_handler_test.rb
@@ -6,7 +6,7 @@ class Folio::Devise::CrossdomainHandlerTest < ActiveSupport::TestCase
   def setup
     super
 
-    create(:sinfin_local_site) # as it it Folio.main_site
+    create(:dummy_site) # as it it Folio.main_site
   end
 
   test "always noop with nil master_site" do
@@ -308,7 +308,7 @@ class Folio::Devise::CrossdomainHandlerTest < ActiveSupport::TestCase
     end
 
     def create_current_site
-      create(:folio_site)
+      create_site
     end
 
     def new_result(request: nil,