+++ test/support/sites_helper.rb
@@ -3,7 +3,7 @@
 module SitesHelper
   private
     def get_any_site
-      ::Folio::Site.first || create(:folio_site)
+      ::Folio::Site.first || create_site
     end
 
     def create_and_host_site(key: nil, attributes: {})
@@ -12,10 +12,15 @@ module SitesHelper
       site
     end
 
-    def create_site(key: nil, attributes: {})
+    def create_site(key: nil, attributes: {}, force: false)
       key_with_fallback = key || Rails.application.config.folio_site_default_test_factory || :folio_site
       factory = FactoryBot.factories[key_with_fallback]
-      factory.send(:class_name).constantize.last || create(key_with_fallback, attributes)
+
+      if force
+        create(key_with_fallback, attributes)
+      else
+        factory.send(:class_name).constantize.last || create(key_with_fallback, attributes)
+      end
     end
 
     def host_site(site)