+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## 2022-06-28
 ### Removed
 - removed `Rails.application.config.folio_console_ability_lambda`. Use `app/overrides/models/folio/ability_override.rb` in your project instead.
+### Changed
+- changed how sites in tests work. Folio expects your project to have a `ApplicationName::Site` class and `Rails.application.folio_site_default_test_factory` set.
 
 ## 2024-06-21
 ### Added